client/src/app/pages/main-page/main-page.component.ts
import { Component, OnDestroy } from '@angular/core';
import { MatDialog, MatDialogRef } from '@angular/material/dialog';
import { Router } from '@angular/router';
import { CreateJoinGameDialogComponent } from '@app/components/create-join-game-dialog/create-join-game-dialog.component';
import { GameMode } from '@common/game-mode';

@Component({
    selector: 'app-main-page',
    templateUrl: './main-page.component.html',
    styleUrls: ['./main-page.component.scss'],
})
export class MainPageComponent implements OnDestroy {
    readonly teamMembers: string[] = [
        'Coralie Brodeur',
        ' Imène Clara Ghazi',
        ' Kylian Chaussoy',
        ' Thibault Demagny',
        ' Younes Benabbou',
        ' Dumitru Zlotea',
    ];
    dialogRef: MatDialogRef<CreateJoinGameDialogComponent>;

    constructor(private readonly router: Router, private dialog: MatDialog) {}

    setGameMode(mode: string) {
        if (mode === GameMode.classicMode) {
            this.router.navigate(['/selection']);
        } else {
            this.dialogRef = this.dialog.open(CreateJoinGameDialogComponent);
        }
    }

    ngOnDestroy() {
        if (this.dialogRef) {
            this.dialogRef.close();
        }
    }
}
